+++ plug-ins/Apple.PHASE/Apple.PHASE_Unity/Assets/CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 - Remove references to some unused packages.
 
 ## [1.2.3] - 2024-12-10
+
+### Added
+- PHASE Spatializer plug-in
+- Listener head-tracking support
+- Personalized HRTF support
+- AVAudioSession interruption handler
 - Add Apache 2.0 license file.
+
+### Changed
+- Initial gain level fix.
+- Updated failing test.
 
 ## [1.2.2] - 2024-11-20
 ### Changed